requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
firebase-admin>=6.4.0
httpx>=0.27.0
pandas>=2.2.0
boto3>=1.34.0
tabulate>=0.9.0
python-multipart>=0.0.9
python-dotenv>=1.0.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
